+++ pkg/account/account.go
@@ -8,11 +8,11 @@ type Account struct {
 }
 
 func (a Account) Sign(unsignedBytes []byte) ([]byte, error) {
-
+	return nil, nil
 }
 
 func (a Account) GetAddr() sdk.AccAddress {
-
+	return nil
 }
 
 func NewAccount() *Account {
